refactor: openapi controller part 1
